LearningRuleType.modifies more explicit
Previously, we used the type of the pre or post object as a clue to
what the learning rule modifies. This makes it more explicit by saying
what signal the learning rule actually modifies. This will make it
easier to implement learning rules that modify other quantities,
like for example `post.encoders`.